setup_sdcard.sh: Add helper is_element_of() function.
Add a simple helper function is_element_of() that will be used to
simplify the validation of various values.  This routine is not meant
to be called directly; instead, it will be used to write much simpler
validation routines to check for acceptable values like a filesystem
type and so on.

Signed-off-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Robert Nelson <robertcnelson@gmail.com>